Extract out surya specific adapters into its own module
